requirements.txt
praw==7.7.1
azure-eventhub==5.11.5
transformers==4.41.2
torch==2.3.1+cpu
numpy==1.26.4